Rename neurobank -> nbank_registry
breaking change
