+++ nbank_registry/errors.py
