+++ nbank_registry/tests.py
@@ -12,8 +12,8 @@ from rest_framework import status
 from rest_framework.test import APITestCase
 
 from django.contrib.auth.models import User
-from neurobank.models import Resource, DataType, Archive, Location
-from neurobank import errors
+from nbank_registry.models import Resource, DataType, Archive, Location
+from nbank_registry import errors
 
 
 class APIAuthTestCase(APITestCase):
